login-signup organization
Tentiva de organizar os signups e login, e também criar o login de motorista.  Problema na comunicação com o Banco de Dados.
