+++ src/Login/Driver.js
@@ -1,25 +1,30 @@
 import React, { useState } from "react";
 import axios from "axios";
-import { StyleSheet, Text, View, TextInput, Alert, Button, Image } from "react-native";
-import {API_REST} from '../api/api'
-import {AUTH_ROUT} from '../api/auth'
-import {PORT} from '../api/port'
+import {
+  StyleSheet,
+  Text,
+  View,
+  TextInput,
+  Alert,
+  Button,
+  Image,
+} from "react-native";
+import { API_REST } from "../api/api";
+import { AUTH_ROUT_DRIVER } from "../api/authdriver";
+import { PORT } from "../api/port";
 
-export default function LogIn({ navigation }) {
+export default function DriverLogIn({ navigation }) {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
 
   function login(email, password) {
-    console.log(API_REST+""+PORT+"/"+AUTH_ROUT)
     axios
-      .post(API_REST+""+PORT+"/"+AUTH_ROUT, {
+      .post(API_REST + "" + PORT + "/" + AUTH_ROUT_DRIVER, {
         email: email,
         password: password,
       })
       .then(function (response) {
         if (response.status == 200) {
-          console.log(response.data);
-          console.log(response.data.token);
           navigation.navigate("Fmenu");
         }
       })
@@ -36,9 +41,9 @@ export default function LogIn({ navigation }) {
         source={{ uri: "https://i.imgur.com/0FltieF.png" }}
       />
       <View style={styles.view}>
-      <Text style={styles.text}>Email</Text>
+        <Text style={styles.text}>Email</Text>
         <TextInput
-        autoCapitalize='none'
+          autoCapitalize="none"
           placeholder="Email"
           value={email}
           onChangeText={setEmail}
@@ -47,7 +52,7 @@ export default function LogIn({ navigation }) {
 
         <Text style={styles.text}>Senha</Text>
         <TextInput
-        autoCapitalize='none'
+          autoCapitalize="none"
           style={styles.fieldInput}
           placeholder="Password"
           value={password}
@@ -94,13 +99,14 @@ const styles = StyleSheet.create({
     width: "100%",
     marginBottom: 25,
   },
-  button:{
-    marginTop:100,
+  button: {
+    marginTop: 100,
     width: 250,
-    height: 75
-  },logo: {
+    height: 75,
+  },
+  logo: {
     width: 240,
     height: 200,
-    marginBottom:5,
+    marginBottom: 5,
   },
 });
