+++ src/api/api.js
@@ -1,2 +1,2 @@
-import axios from  'axios'
+export const API_REST = "http://192.168.0.6/:"
