+++ src/api/authdriver.js
@@ -1,0 +1,1 @@
+export const AUTH_ROUT_DRIVER = 'api/authdriver'
