+++ src/api/authuser.js
@@ -1,0 +1,1 @@
+export const AUTH_ROUT_USER = 'api/authuser'
